requirements.txt
requests
torch
pandas
numpy
scikit-learn
matplotlib
joblib
tqdm
ccxt
python-dotenv
fastapi
uvicorn
jinja2
python-multipart